+++ Challenge.lua
@@ -30,14 +30,6 @@ function MyDungeonsBook:InitNewDungeonChallenge(id)
 		id = id,
 		challengeInfo = {},
 		gameInfo = {},
-		details = {
-			exists = false,
-			player = {},
-			party1 = {},
-			party2 = {},
-			party3 = {},
-			party4 = {}
-		},
 		encounters = {},
 		players = {
 			player = {},
@@ -89,97 +81,6 @@ function MyDungeonsBook:ParseUnitInfoWithWowApi(unit)
 		talents = {},
 		misc = {}
 	};
-end
-
---[[--
-Parse info from Details addon about all party members DPS, HPS etc.
-
-**It must be called only when whole party is together (and anyone didn't leave it).**
-
-@return[type=table] details for all party members
-]]
-function MyDungeonsBook:ParseInfoFromDetailsAddon()
-	if(not IsAddOnLoaded("Details")) then
-		self:LogPrint("Addon Details is not loaded!");
-		return {exists = false};
-	end
-	local details = {
-		exists = true
-	};
-	local combat = Details:GetCombat("overall");
-	if (not combat) then
-		self:LogPrint("Combat 'overall' not exists in the Details!");
-		return {exists = false};
-	end
-	for _, unit in pairs(self:GetPartyRoster()) do
-		local name, realm = UnitName(unit);
-		local detailsUnitName = name;
-		if (realm) then
-			detailsUnitName = string.format("%s-%s", detailsUnitName, realm);
-		end
-		details[detailsUnitName] = self:ParseUnitInfoFromDetailsAddon(detailsUnitName or "NOT FOUND");
-		self:DebugPrint(string.format("Details info for %s is saved", unit));
-	end
-	return details;
-end
-
---[[--
-Get info from Details addon for a single party member.
-
-It can be called in any time while you didn't reset Details.
-
-@param[type=string] detailsUnitName
-@return[type=table] details for single a party member
-]]
-function MyDungeonsBook:ParseUnitInfoFromDetailsAddon(detailsUnitName)
-	local details = {};
-	local combat = Details:GetCombat("overall");
-	local damageActor = combat:GetActor(DETAILS_ATTRIBUTE_DAMAGE, detailsUnitName);
-		if (damageActor) then
-			local totalDamage = (damageActor and damageActor.total) or 0;
-			local activeDps, effectiveDps;
-			if (totalDamage) then
-				activeDps = totalDamage / combat:Tempo();
-				effectiveDps = totalDamage / combat:GetCombatTime();
-			else
-				activeDps = "-";
-				effectiveDps = "-";
-			end
-			details.totalDamage = totalDamage;
-			details.activeDps = activeDps;
-			details.effectiveDps = effectiveDps;
-		else
-			self:DebugPrint(string.format("%s actor not found for unit %s", "DAMAGE", detailsUnitName));
-		end
-
-		local healActor = combat:GetActor(DETAILS_ATTRIBUTE_HEAL, detailsUnitName);
-		if (healActor) then
-			local totalHeal = (healActor and healActor.total) or 0;
-			local activeHps, effectiveHps;
-			if (totalHeal) then
-				activeHps = totalHeal / combat:Tempo();
-				effectiveHps = totalHeal / combat:GetCombatTime();
-			else
-				activeHps = "-";
-				effectiveHps = "-";
-			end
-			details.totalHeal = totalHeal;
-			details.activeHps = activeHps;
-			details.effectiveHps = effectiveHps;
-		else
-			self:DebugPrint(string.format("%s actor not found for unit %s", "HEAL", detailsUnitName));
-		end
-
-		local miscActor = combat:GetActor(DETAILS_ATTRIBUTE_MISC, detailsUnitName);
-		if (miscActor) then
-			local interrupt = miscActor.interrupt;
-			local dispell = miscActor.dispell;
-			details.interrupt = interrupt;
-			details.dispell = dispell;
-		else
-			self:DebugPrint(string.format("%s actor not found for unit %s", "MISC", detailsUnitName));
-		end
-		return details;
 end
 
 --[[--