+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ﻿### v1.8.0
 
+* [#16](https://github.com/onechiporenko/my-dungeons-book/issues/16) Reformat "Details"-tab to "Summary"-tab
 * [#82](https://github.com/onechiporenko/my-dungeons-book/issues/82) MDB window should be closable via ESC pressing
 * [#83](https://github.com/onechiporenko/my-dungeons-book/issues/83) Add percentage column for tables "Heal by Spell" and "All Damage Done"